requirements.txt
pyatv
pytz
requests
python-dateutil
protobuf
aiohttp
cchardet
aiodns
PyYAML
lxml
async_retrying